Implement entity linking, untested yet

+++ src/main/java/org/kddm2/search/entity/EntityWeightingTFIDF.java
@@ -3,22 +3,18 @@ package org.kddm2.search.entity;
 import org.kddm2.indexing.IndexStatsHelper;
 import org.kddm2.indexing.IndexTermStats;
 
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-
 public class EntityWeightingTFIDF extends EntityWeightingAlgorithm {
     public EntityWeightingTFIDF(IndexStatsHelper indexHelper, EntityTools entityTools) {
         super(indexHelper, entityTools);
     }
 
     @Override
-    protected double getWeightForCandidate(EntityCandidate candidate, int occurenceCount) {
+    protected float getWeightForCandidate(EntityCandidate candidate, int occurenceCount) {
         IndexTermStats statsForDictTerm = indexHelper.getStatsForDictTerm(candidate.getCandidateText());
         if(statsForDictTerm.getCountOccurenceDocuments() > 1) {
-            return occurenceCount / Math.log(statsForDictTerm.getCountOccurenceDocuments());
+            return (float) (occurenceCount / Math.log(statsForDictTerm.getCountOccurenceDocuments()));
         } else {
-            return occurenceCount * Math.E;
+            return (float) (occurenceCount * Math.E);
         }
     }
 }
